Premium Analytics: rename over-time report controllers to Snake_Case (WPCS)
