+++ projects/packages/premium-analytics/src/Reports/Export/Exports/class-orders-fulfilled-over-time-controller.php
@@ -11,7 +11,7 @@ namespace Automattic\Jetpack\PremiumAnalytics\Reports\Export\Exports;
 
 defined( 'ABSPATH' ) || exit;
 
-use Automattic\Jetpack\PremiumAnalytics\Reports\Export\AbstractCSVReportController;
+use Automattic\Jetpack\PremiumAnalytics\Reports\Export\Abstract_Csv_Report_Controller;
 
 /**
  * Orders Fulfilled Over Time CSV Export Controller.
@@ -21,7 +21,7 @@ use Automattic\Jetpack\PremiumAnalytics\Reports\Export\AbstractCSVReportControll
  *
  * @since $$next-version$$
  */
-class OrdersFulfilledOverTimeController extends AbstractCSVReportController {
+class Orders_Fulfilled_Over_Time_Controller extends Abstract_Csv_Report_Controller {
 
 	/**
 	 * Get the report key for this controller.
