+++ projects/packages/premium-analytics/src/Reports/Export/Exports/class-taxes-over-time-controller.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * REST API Reports Visitors Over Time controller class.
+ * REST API Reports Taxes Over Time controller class.
  *
  * @package Automattic\Jetpack\PremiumAnalytics\Reports\Export\Exports
  */
@@ -11,17 +11,17 @@ namespace Automattic\Jetpack\PremiumAnalytics\Reports\Export\Exports;
 
 defined( 'ABSPATH' ) || exit;
 
-use Automattic\Jetpack\PremiumAnalytics\Reports\Export\AbstractCSVReportController;
+use Automattic\Jetpack\PremiumAnalytics\Reports\Export\Abstract_Csv_Report_Controller;
 
 /**
- * Visitors Over Time CSV Export Controller.
+ * Taxes Over Time CSV Export Controller.
  *
- * Handles CSV exports for the Visitors Over Time report, supporting both
+ * Handles CSV exports for the Taxes Over Time report, supporting both
  * single interval and comparison interval data.
  *
  * @since $$next-version$$
  */
-class VisitorsOverTimeController extends AbstractCSVReportController {
+class Taxes_Over_Time_Controller extends Abstract_Csv_Report_Controller {
 
 	/**
 	 * Get the report key for this controller.
@@ -29,7 +29,7 @@ class VisitorsOverTimeController extends AbstractCSVReportController {
 	 * @return string The report key.
 	 */
 	public function get_report_key(): string {
-		return 'visitorsovertime';
+		return 'taxesovertime';
 	}
 
 	/**
@@ -38,7 +38,7 @@ class VisitorsOverTimeController extends AbstractCSVReportController {
 	 * @return string The report label.
 	 */
 	public function get_report_label(): string {
-		return __( 'Visitors Over Time', 'jetpack-premium-analytics' );
+		return __( 'Taxes Over Time', 'jetpack-premium-analytics' );
 	}
 
 	/**
@@ -47,7 +47,7 @@ class VisitorsOverTimeController extends AbstractCSVReportController {
 	 * @return string The data endpoint.
 	 */
 	public function get_data_endpoint(): string {
-		return 'reports/sessions/by-date';
+		return 'reports/taxes/by-date';
 	}
 
 	/**
@@ -59,7 +59,10 @@ class VisitorsOverTimeController extends AbstractCSVReportController {
 	public function get_column_headers( ?string $interval = null ): array {
 		return array(
 			'time_interval' => $this->get_interval_label( $interval ),
-			'visitors'      => __( 'Visitors', 'jetpack-premium-analytics' ),
+			'orders_count'  => __( 'Orders', 'jetpack-premium-analytics' ),
+			'order_tax'     => __( 'Order tax', 'jetpack-premium-analytics' ),
+			'shipping_tax'  => __( 'Shipping tax', 'jetpack-premium-analytics' ),
+			'total_tax'     => __( 'Total tax', 'jetpack-premium-analytics' ),
 		);
 	}
 
@@ -70,7 +73,10 @@ class VisitorsOverTimeController extends AbstractCSVReportController {
 	 */
 	public function get_default_values(): array {
 		return array(
-			'visitors' => 0,
+			'orders_count' => 0,
+			'order_tax'    => 0,
+			'shipping_tax' => 0,
+			'total_tax'    => 0,
 		);
 	}
 
@@ -85,7 +91,21 @@ class VisitorsOverTimeController extends AbstractCSVReportController {
 		$defaults = $this->get_default_values();
 		return array(
 			'time_interval' => $this->format_time_interval( $item, $interval ),
-			'visitors'      => $item['visitors'] ?? $defaults['visitors'],
+			'orders_count'  => $item['orders_count'] ?? $defaults['orders_count'],
+			'order_tax'     => self::format_amount( $item['order_tax'] ?? $defaults['order_tax'] ),
+			'shipping_tax'  => self::format_amount( $item['shipping_tax'] ?? $defaults['shipping_tax'] ),
+			'total_tax'     => self::format_amount( $item['total_tax'] ?? $defaults['total_tax'] ),
+		);
+	}
+
+	/**
+	 * Get additional request parameters for data fetching.
+	 *
+	 * @return array Additional parameters to include in data requests.
+	 */
+	public function get_additional_params(): array {
+		return array(
+			'date_type' => self::DEFAULT_DATE_TYPE,
 		);
 	}
 
@@ -98,7 +118,10 @@ class VisitorsOverTimeController extends AbstractCSVReportController {
 		// time_interval, date_start, and date_end are always returned by the
 		// API for time-series endpoints and do not need to be requested.
 		return array(
-			'visitors',
+			'orders_count',
+			'order_tax',
+			'shipping_tax',
+			'total_tax',
 		);
 	}
 }
